# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Auction System API",
        "title": "Auction System API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/api/auction": {
            "post": {
                "description": "创建拍卖品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "拍卖品"
                ],
                "summary": "创建拍卖品",
                "parameters": [
                    {
                        "description": "创建拍卖品请求参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.CreateAuctionRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/kernal.SuccessResult"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/auction/bid": {
            "post": {
                "description": "创建出价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "出价"
                ],
                "summary": "创建出价",
                "parameters": [
                    {
                        "description": "创建出价请求参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.CreateBidRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "$ref": "#/definitions/kernal.SuccessResult"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/auction/bid/higest": {
            "get": {
                "description": "获取拍卖品的最高出价",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "出价"
                ],
                "summary": "获取拍卖品的最高出价",
                "parameters": [
                    {
                        "description": "获取拍卖品的最高出价请求参数",
                        "name": "req",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/rest.GetHigestBidRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/kernal.SuccessResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/application.BidBriefDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    }
                }
            }
        },
        "/api/auction/latest": {
            "get": {
                "description": "获取最新的拍卖品",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "拍卖品"
                ],
                "summary": "获取最新的拍卖品",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "页码",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "每页数量",
                        "name": "size",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/kernal.SuccessResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/application.AuctionBriefDTO"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "登录获取访问令牌和刷新令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "登录请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.LoginRequestSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.LoginResponseSchema"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    }
                }
            }
        },
        "/auth/refresh_token": {
            "post": {
                "description": "使用刷新令牌刷新访问令牌",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Refresh token",
                "parameters": [
                    {
                        "description": "刷新令牌请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.RefreshTokenRequestSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/kernal.SuccessResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/auth.RefreshTokenResponseSchema"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    }
                }
            }
        },
        "/auth/signup": {
            "post": {
                "description": "注册新用户",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "注册",
                "parameters": [
                    {
                        "description": "注册请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/auth.SignupRequestSchema"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/auth.SignupResponseSchema"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    }
                }
            }
        },
        "/ws": {
            "get": {
                "description": "处理ws连接",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "ws"
                ],
                "summary": "处理ws连接",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "拍卖id",
                        "name": "auction_id",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/kernal.ErrorResult"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "application.AuctionBriefDTO": {
            "type": "object",
            "properties": {
                "end_time": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "init_price": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "integer"
                },
                "status": {
                    "$ref": "#/definitions/shared.AuctionStatus"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "application.BidBriefDTO": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "auth.LoginRequestSchema": {
            "type": "object",
            "required": [
                "name",
                "password"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.LoginResponseSchema": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "access_token_expire_time": {
                    "type": "string"
                },
                "access_token_type": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "refresh_token_expire_time": {
                    "type": "string"
                },
                "refresh_token_type": {
                    "type": "string"
                }
            }
        },
        "auth.RefreshTokenRequestSchema": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "auth.RefreshTokenResponseSchema": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "access_token_expire_time": {
                    "type": "string"
                },
                "access_token_type": {
                    "type": "string"
                },
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                },
                "refresh_token": {
                    "type": "string"
                },
                "refresh_token_expire_time": {
                    "type": "string"
                },
                "refresh_token_type": {
                    "type": "string"
                }
            }
        },
        "auth.SignupRequestSchema": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "auth.SignupResponseSchema": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "kernal.ErrorResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "kernal.SuccessResult": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "rest.CreateAuctionRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "end_time": {
                    "type": "integer"
                },
                "init_price": {
                    "type": "integer"
                },
                "start_time": {
                    "type": "integer"
                },
                "step_price": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "rest.CreateBidRequest": {
            "type": "object",
            "properties": {
                "auction_id": {
                    "type": "integer"
                },
                "price": {
                    "type": "integer"
                }
            }
        },
        "rest.GetHigestBidRequest": {
            "type": "object",
            "required": [
                "auction_id"
            ],
            "properties": {
                "auction_id": {
                    "type": "integer"
                }
            }
        },
        "shared.AuctionStatus": {
            "type": "integer",
            "enum": [
                1,
                2,
                3
            ],
            "x-enum-varnames": [
                "AuctionStatusRunning",
                "AuctionStatusEnded",
                "AuctionStatusWaiting"
            ]
        }
    }
}